refactor JetpacConnectionkWebView and its client to display error screen when error happens during request
